requirements.txt
azure-cognitiveservices-vision-computervision==0.6.0
azure-common==1.1.25
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
cymem==2.0.3
idna==2.10
isodate==0.6.0
msrest==0.6.19
murmurhash==1.0.2
numpy==1.19.2
oauthlib==3.1.0
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
spacy==2.3.2
srsly==1.0.2
thinc==7.4.1
tqdm==4.49.0
urllib3==1.25.10
wasabi==0.8.0
https://med7.s3.eu-west-2.amazonaws.com/en_core_med7_lg.tar.gz